src/cloudfunction/batch_train/requirements.txt
google-cloud-dataproc==2.2.0